+++ Jobsy.Web/wwwroot/js/maps-loader.js
@@ -13,7 +13,7 @@ window.jobsyMaps = (function () {
         "/js/jobsyMapLibre.min.js?v=20260822-r195"
     ];
     var discoveryScripts = [
-        "/js/jobMap.min.js?v=20260822-r202"
+        "/js/jobMap.min.js?v=20260822-r205"
     ];
     var detailScripts = [
         "/js/vacancyDetailMap.min.js?v=20260822-r195"